move validator functions back to validators page
